add: new drawer component for mobile navigation
