+++ app/components/app/Header.vue
@@ -1,59 +1,18 @@
 <script setup lang="ts">
-import { ref, reactive, onMounted, watch } from 'vue'
-import { useRoute } from 'vue-router'
+import { ref } from 'vue'
 import { colorMap } from '~/utils/colorMap'
 import { animations } from '~/utils/animations'
+import { useNavigationLinks } from '~/composables/useNavigationLinks'
 
 const props = defineProps<{ color: string }>()
 const active = ref('Inicio')
-const isOpen = ref(false)
-const route = useRoute()
-
-const navigationLinks = reactive([
-  {
-    name: "Inicio",
-    link: "/",
-  },
-  {
-    name: "Nosotros",
-    link: "/nosotros",
-  },
-  {
-    name: "Servicios",
-    link: "/servicios",
-  },
-  {
-    name: "Portafolio",
-    link: "/portafolio",
-  },
-  {
-    name: "Blog",
-    link: "/blog",
-  },
-])
-
-function drawer() {
-  isOpen.value = !isOpen.value
-}
+const { navigationLinks } = useNavigationLinks()
 
 function setActive(linkName: string) {
   active.value = linkName
 }
 
-watch(isOpen, (val) => {
-  if (process.client) {
-    if (val) document.body.style.setProperty("overflow", "hidden")
-    else document.body.style.removeProperty("overflow")
-  }
-}, { immediate: true })
-
-onMounted(() => {
-  document.addEventListener("keydown", (e: KeyboardEvent) => {
-    if (e.key === "Escape" && isOpen.value) isOpen.value = false
-  })
-})
 </script>
-
 
 <template>
   <header class="fixed top-0 w-full bg-transparent px-6 pt-8 md:px-9 lg:px-12 lg:pt-12 transition-colors"
@@ -76,19 +35,6 @@ onMounted(() => {
         </div>
       </Motion>
 
-      <!-- Mobile toggle -->
-      <Motion :initial="animations.header.initial" :animate="animations.header.animate"
-        :transition="{ ...animations.header.transition }">
-        <div class="md:hidden">
-          <button @click="isOpen = !isOpen">
-            <svg class="h-8 w-8 fill-current text-black" fill="none" stroke-linecap="round" stroke-linejoin="round"
-              stroke-width="2" viewBox="0 0 24 24" stroke="currentColor">
-              <path d="M4 6h16M4 12h16M4 18h16"></path>
-            </svg>
-          </button>
-        </div>
-      </Motion>
-
       <!-- Navbar (Desktop) -->
       <div class="hidden lg:justify-end md:flex">
         <Motion :initial="animations.header.initial" :animate="animations.header.animate"
@@ -104,57 +50,6 @@ onMounted(() => {
           </ul>
         </Motion>
       </div>
-
-
-      <transition enter-class="opacity-0" enter-active-class="ease-out transition-medium" enter-to-class="opacity-100"
-        leave-class="opacity-100" leave-active-class="ease-out transition-medium" leave-to-class="opacity-0">
-        <div @keydown.esc="isOpen = false" v-show="isOpen" class="z-10 fixed inset-0 transition-opacity">
-          <div @click="isOpen = false" class="absolute inset-0 bg-black opacity-50" tabindex="0"></div>
-        </div>
-      </transition>
-
-      <!-- Menu (Móvil) -->
-      <aside
-        class="flex flex-col pt-8 pl-6 pr-4 transform top-0 right-0 w-2/5 bg-eerieBlack fixed h-full overflow-auto ease-in-out transition-all duration-300 z-30"
-        :class="isOpen ? 'translate-x-0' : 'translate-x-full'">
-        <div class="close">
-          <button class="flex w-full justify-end" @click="isOpen = false">
-            <svg width="23" height="23" viewBox="0 0 20 20" fill="none" xmlns="http://www.w3.org/2000/svg">
-              <circle cx="10" cy="10" r="9.60805" stroke="#007FFF" stroke-width="0.783908" />
-              <path
-                d="M12.5078 5.13281H13.6797L10.5625 9.32812L13.8203 13.6875H12.6484L10 10.0547L7.35156 13.6875H6.17969L9.4375 9.30469L6.34375 5.13281H7.51562L10 8.57812L12.5078 5.13281Z"
-                fill="#007FFF" />
-            </svg>
-          </button>
-        </div>
-
-        <ul class="flex flex-col gap-5 text-columbiaBlue hover:bottom-1">
-          <li v-for="(item, index) in navigationLinks" :key="index">
-            <NuxtLink :to="item.link" :class="[
-              'flex justify-start focus:outline-none',
-              $route.path === item.link ? 'border-b-2 border-azure text-azure' : '',
-            ]" @click="isOpen = false">
-              {{ item.name }}
-            </NuxtLink>
-          </li>
-        </ul>
-        <!-- Social Media Icons (Mobile) -->
-        <div class="flex gap-1 h-9 w-full justify-start">
-          <a href="https://maps.app.goo.gl/rzpD2iuE3SB6jsXEA" target="_blank" rel="noopener noreferrer">
-            <Icon name="material-symbols:location-on" class="text-xl md:text-2xl lg:text-2xl" />
-          </a>
-          <a href="https://www.instagram.com/novanetstudio/" target="_blank" rel="noopener noreferrer">
-            <Icon name="fa6-brands:instagram" class="text-xl md:text-2xl lg:text-2xl" />
-          </a>
-          <a href="https://wa.me/message/QA5DVPVFSXYCJ1" target="_blank" rel="noopener noreferrer">
-            <Icon name="fa6-brands:whatsapp" class="text-xl md:text-2xl lg:text-2xl" />
-          </a>
-          <a href="https://www.facebook.com/NovanetStudio" target="_blank" rel="noopener noreferrer">
-            <Icon name="fa6-brands:facebook" class="text-xl md:text-2xl lg:text-2xl" />
-          </a>
-        </div>
-
-      </aside>
     </nav>
   </header>
 </template>
