+++ app/layouts/default.vue
@@ -2,7 +2,6 @@
 import { useSectionColor } from '~/composables/useSectionColor'
 import { useRoute } from 'vue-router'
 import { watch, nextTick } from 'vue'
-
 const { currentColor, initObserver } = useSectionColor()
 const route = useRoute()
 
@@ -15,9 +14,10 @@ watch(
 )
 </script>
 <template>
-   <AppHeader :color="currentColor" />
-   <main>
-      <slot />
-   </main>
-   <AppFooter :color="currentColor" />
+  <AppHeader :color="currentColor" />
+  <AppDrawer :color="currentColor" />
+  <main>
+    <slot />
+  </main>
+  <AppFooter :color="currentColor" />
 </template>
